test: add demo table as sub routine
